+++ src/main/background.js
@@ -1,0 +1,3 @@
+import './omnibar'
+import './menus-and-shortcuts'
+import './trigger-capture'
