+++ src/main/omnibar.js
@@ -2,8 +2,7 @@ import debounce from 'lodash/fp/debounce'
 import fromPairs from 'lodash/fp/fromPairs'
 import escapeHtml from 'lodash/fp/escape'
 
-import { getTimestamp } from 'src/page-storage'
-import { filterPagesByQuery } from 'src/search'
+import { getTimestamp, filterPagesByQuery } from 'src/page-storage'
 import { hrefForLocalPage } from 'src/local-page'
 import niceTime from 'src/util/nice-time'
