+++ src/manifest.json
@@ -6,7 +6,7 @@
         "96": "assets/webmemex-96.png"
     },
     "background": {
-        "scripts": ["lib/browser-polyfill.js", "background.js"]
+        "scripts": ["lib/browser-polyfill.js", "main/background.js"]
     },
     "content_scripts": [{
         "matches": ["<all_urls>"],
