+++ src/page-storage/delete-page.js
