+++ src/page-storage/index.js
@@ -1,6 +1,12 @@
 import docuri from 'docuri'
 import randomString from 'src/util/random-string'
 
+export * from './create-page'
+export * from './delete-page'
+export * from './download-page'
+export * from './get-pages'
+export * from './text-search'
+
 export const pageKeyPrefix = 'page/'
 
 export const convertPageDocId = docuri.route(`${pageKeyPrefix}:timestamp/:nonce`)
